+++ hrepr/__init__.py
@@ -1,7 +1,11 @@
 """Extensible HTML representation for Python objects."""
 
+from . import h
 from .core import Config, Hrepr, HreprState, Interface, StdHrepr  # noqa: F401
 from .h import HTML, H, HType, Tag  # noqa: F401
+from .hgen import standard_html
+
+h.standard_html = standard_html
 
 config_defaults = {
     "string_cutoff": 20,
