+++ hrepr/h.py
@@ -1,48 +1,17 @@
-import json
 import os.path
-from html import escape
-from itertools import count
-from types import FunctionType, GeneratorType, MethodType, SimpleNamespace
-from typing import Sequence, Union
+from types import GeneratorType
+from typing import Union
 
-from ovld import OvldMC, ovld
+from ovld import ovld
 
-from .textgen import Breakable, Context, Sequence, Text, join
+from .textgen import Context
 
 # CSS for hrepr
 styledir = f"{os.path.dirname(__file__)}/style"
 css_nbreset = open(f"{styledir}/nbreset.css", encoding="utf-8").read()
 
-
-# These tags are self-closing and cannot have children.
-_void_tags = {
-    "area",
-    "base",
-    "br",
-    "col",
-    "command",
-    "embed",
-    "hr",
-    "img",
-    "input",
-    "keygen",
-    "link",
-    "meta",
-    "param",
-    "source",
-    "track",
-    "wbr",
-}
-
-
-# We do not escape string children for these tags.
-_raw_tags = {"raw", "script", "style"}
-
-
-# These tags do not correspond to real HTML tags. They cannot have
-# attributes, and their children are simply concatenated and inlined
-# in the parent.
-_virtual_tags = {None, "inline", "raw"}
+# Used by __str__, set by __init__ to avoid a circular dependency
+standard_html = None
 
 
 def flatten(seq):
@@ -284,308 +253,3 @@ class HTML:
 
 H = HTML(tag_class=Tag, instantiate=True)
 HType = HTML(tag_class=Tag, instantiate=False)
-
-
-class HTMLGenerator(metaclass=OvldMC):
-    def __init__(self, rules):
-        self.rules = rules
-
-    def register(self, rule, fn=None):
-        def reg(fn):
-            self.rules[rule] = fn
-            return fn
-
-        if fn is None:
-            return reg
-        else:
-            return reg(fn)
-
-    def _default_tag(self, node, workspace, default):
-        return default(node, workspace)
-
-    def _default_attr(self, node, workspace, attr, value, default):
-        return default(node, workspace, attr, value)
-
-    def default_tag(self, node, workspace):
-        workspace.open = node.name
-        workspace.close = node.name
-
-    def default_attr(self, node, workspace, attr, value):
-        if value is False:
-            return
-        elif value is True:
-            workspace.attributes[attr] = value
-        elif isinstance(value, (tuple, set, frozenset)):
-            workspace.attributes[attr] = " ".join(escape(cls) for cls in value)
-        elif isinstance(value, str):
-            workspace.attributes[attr] = escape(value)
-        else:
-            workspace.attributes[attr] = escape(json.dumps(value))
-
-    @ovld
-    def process(self, node: str):
-        return node
-
-    @ovld
-    def process(self, node: object):
-        return str(node)
-
-    @ovld
-    def process(self, node: Tag):
-        workspace = SimpleNamespace(
-            open=None,
-            close=None,
-            attributes={},
-            children=[],
-            resources=[],
-            extra=[],
-            escape_children=True,
-        )
-
-        workspace.children = [
-            self.process(child) for child in iterate_children(node.children)
-        ]
-        for child in workspace.children:
-            if getattr(child, "extra", None):
-                workspace.extra += child.extra
-                child.extra = []
-
-        tag_rule = self.rules.get(f"tag:{node.name}", self._default_tag)
-        if tag_rule(node, workspace, self.default_tag) is False:
-            pass
-        else:
-            for k, v in node.attributes.items():
-                rule = self.rules.get(f"attr:{k}", self._default_attr)
-                if rule(node, workspace, k, v, self.default_attr) is False:
-                    break
-
-        return workspace
-
-    def _text_parts(self, workspace):
-        def convert_child(c):
-            if isinstance(c, str):
-                if workspace.escape_children:
-                    return escape(str(c))
-                else:
-                    return c
-            else:
-                return self._text_parts(c)
-
-        attr = "".join(
-            f" {k}" if v is True else f' {k}="{v}"'
-            for k, v in workspace.attributes.items()
-        )
-
-        children = list(map(convert_child, workspace.children))
-
-        if workspace.open:
-            if workspace.close:
-                return Breakable(
-                    start=f"<{workspace.open}{attr}>",
-                    body=children,
-                    end=f"</{workspace.open}>",
-                )
-            else:
-                assert not children
-                return Text(f"<{workspace.open}{attr} />")
-        else:
-            return Breakable(start=None, body=children, end=None)
-
-    def generate(self, node):
-        ws = self.process(node)
-        to_process = [ws, *[self.process(x) for x in ws.extra]]
-        parts = [self._text_parts(x) for x in to_process]
-        if len(parts) == 1:
-            return parts[0]
-        else:
-            return Breakable(start=None, body=parts, end=None)
-
-
-standard_html = HTMLGenerator({})
-
-
-##########################
-# Special JSON generator #
-##########################
-
-
-_type = type
-
-
-@ovld
-def dump(self, d: dict):
-    return Breakable(
-        start="{",
-        body=join(
-            [Sequence(self(k), ": ", self(v)) for k, v in d.items()], sep=", "
-        ),
-        end="}",
-    )
-
-
-@ovld
-def dump(self, seq: Union[list, tuple]):
-    return Breakable(start="[", body=join(map(self, seq), sep=", "), end="]",)
-
-
-@ovld
-def dump(self, x: Union[int, float, str, bool, _type(None)]):
-    return Text(json.dumps(x))
-
-
-@ovld
-def dump(self, fn: Union[FunctionType, MethodType]):
-    return self(None)
-
-
-@ovld
-def dump(self, t: HType.self):
-    return f"self"
-
-
-@ovld
-def dump(self, t: Tag):
-    tag_id = t.get_attribute("id")
-    if not tag_id:
-        raise ValueError(f"Cannot embed <{t.name}> element without an id.")
-    return f"document.getElementById('{tag_id}')"
-
-
-@ovld
-def dump(self, d: object):
-    raise TypeError(
-        f"Objects of type {type(d).__name__} cannot be JSON-serialized."
-    )
-
-
-def dumps(obj, **fmt):
-    return dump(obj).to_string(**fmt)
-
-
-######################################
-# Handlers for void/raw/special tags #
-######################################
-
-
-def raw_tag(node, workspace, default):
-    workspace.escape_children = False
-    return default(node, workspace)
-
-
-def void_tag(node, workspace, default):
-    assert not workspace.children
-    workspace.open = node.name
-    workspace.close = None
-
-
-def virtual_tag(node, workspace, default):
-    return
-
-
-def raw_virtual_tag(node, workspace, default):
-    workspace.escape_children = False
-
-
-# These tags are self-closing and cannot have children.
-_void_tags = {
-    "area",
-    "base",
-    "br",
-    "col",
-    "command",
-    "embed",
-    "hr",
-    "img",
-    "input",
-    "keygen",
-    "link",
-    "meta",
-    "param",
-    "source",
-    "track",
-    "wbr",
-}
-
-for t in _void_tags:
-    standard_html.register(f"tag:{t}", void_tag)
-
-
-# We do not escape string children for these tags.
-_raw_tags = {"script", "style"}
-
-for t in _raw_tags:
-    standard_html.register(f"tag:{t}", raw_tag)
-
-
-# These tags do not correspond to real HTML tags. They cannot have
-# attributes, and their children are simply concatenated and inlined
-# in the parent.
-_virtual_tags = {None, "inline"}
-
-for t in _virtual_tags:
-    standard_html.register(f"tag:{t}", virtual_tag)
-
-
-_raw_virtual_tags = {"raw"}
-
-for t in _raw_virtual_tags:
-    standard_html.register(f"tag:{t}", raw_virtual_tag)
-
-
-###################################
-# Handlers for special attributes #
-###################################
-
-
-constructor_template_script = """{imp}
-const self = document.getElementById('{node_id}');
-const arglist = {arguments};
-const obj = new {symbol}(...arglist);
-window.${node_id} = obj;
-"""
-
-
-id_counter = count()
-
-
-@standard_html.register("attr:constructor")
-def constructor_attribute(node, workspace, key, value, default):
-    if "id" in node.attributes:
-        node_id = node.attributes["id"]
-    else:
-        node_id = workspace.attributes["id"] = f"$hrepr${next(id_counter)}"
-
-    module = value.get("module", None)
-
-    if module:
-        symbol = value.get("symbol", None)
-        if symbol is None:
-            symbol = "constructor"
-            imp = f"import constructor from '{module}';"
-        else:
-            imp = f"import {{{symbol}}} from '{module}';"
-    else:
-        symbol = value["symbol"]
-        imp = ""
-
-    if "options" in value:
-        assert "arguments" not in value
-        arguments = [H.self(), value["options"]]
-    elif "arguments" in value:
-        arguments = value["arguments"]
-        if isinstance(arguments, dict):
-            arguments = [arguments]
-    else:
-        arguments = [H.self()]
-
-    sc = H.script(
-        constructor_template_script.format(
-            imp=imp,
-            symbol=symbol,
-            node_id=node_id,
-            arguments=dumps(arguments, tabsize=4, max_col=80),
-        ),
-        type="module",
-    )
-
-    workspace.extra.append(sc)
