+++ tests/test_hgen/test_constructor_global_symbol.html
