+++ tests/test_hgen/test_constructor_multiple.html
@@ -11,14 +11,14 @@
             }
         }
     }
-    </script></head><body><div><h2>The buttons should increment by 2 and 3 respectively.</h2><button style="width:100px;" id="$hrepr$0">ERROR!</button><button style="width:100px;" id="$hrepr$1">ERROR!</button></div></body></html><script type="module">
-const self = document.getElementById('$hrepr$0');
+    </script></head><body><div><h2>The buttons should increment by 2 and 3 respectively.</h2><button style="width:100px;" id="$hrepr$1">ERROR!</button><button style="width:100px;" id="$hrepr$2">ERROR!</button></div></body></html><script type="module">
+const self = document.getElementById('$hrepr$1');
 const arglist = [self, {"increment": 2}];
 const obj = new Counter(...arglist);
-window.$$hrepr$0 = obj;
+window.$$hrepr$1 = obj;
 </script><script type="module">
-const self = document.getElementById('$hrepr$1');
+const self = document.getElementById('$hrepr$2');
 const arglist = [self, {"increment": 3}];
 const obj = new Counter(...arglist);
-window.$$hrepr$1 = obj;
+window.$$hrepr$2 = obj;
 </script>
